Vit_model/requirements.txt
torch
torchvision
pandas
scikit-learn
pillow
numpy
tqdm
